requirements.txt
Flask
mysql-connector-python
pdfplumber
pytesseract
pdf2image
pandas
matplotlib
